==> 1. ETL.ipynb <==
LEER LOS ARCHIVOS

#IMPORTAR LAS LIBRERIAS
import json
import pandas as pd
import ast

#LEER EL PRIMER ARCHIVO DE JSON
rows=[]
with open("australian_user_reviews.json",encoding="MacRoman")as file:
    for line in file.readlines():
        rows.append(ast.literal_eval(line))
game_reviews=pd.DataFrame(rows)

#LEER EL SEGUNDO ARCHIVO
rows=[]
with open("australian_users_items.json",encoding="MacRoman")as file:
    for line in file.readlines():
        rows.append(ast.literal_eval(line))
game_items=pd.DataFrame(rows)

#LEER EL TERCER ARCHIVO
steam=pd.read_json("output_steam_games.json", lines=True)

TRABAJAR EL PRIMER ARCHIVO

#USAR EXPLODE PARA SEPARAR LOS DICCIONARIOS EN VARIAS FILAS
prueba1=game_reviews.explode("reviews")

#VER EL CONTENIDO DE LA PRIMERA FILA DE LA COLUMNA REVIEWS
game_reviews.iloc[0,2]

#CONVERIR LA COLUMNA DE DICCIONARIOS EN COLUMNAS DE UN DATAFRAME
pd1=prueba1["reviews"].apply(pd.Series)

# UNIR LAS COLUMNAS Y ELIMINAR LA COLUMNA CON LOS DICCIONARIOS
final=pd.concat([game_reviews,pd1], axis="columns")

#eliminar la columna reviews
reviews=final.drop(["reviews"], axis=1)

#HABIA UNA FILA QUE SE PUSO AL FINAL CON EL VALOR DE O, Y NO SE VA A INCLUIR EN EL DATASET
reviews=reviews.iloc[:,0:9]

reviews

#PASAR EL DATAFRAME A UN ARCHIVO CSV
reviews.to_csv("game_reviews.csv", index=False)

TRABAJAR EL SEGUNDO ARCHIVO

#ELIMINAR LAS NAN DEL ARCHIVO
game_items.dropna(thresh=5, inplace=True)

#VER QUE HAY EN LA PRIMERA FILA Y EN LA COLUMNA DE ITEMS
game_items["items"][0]

#SEPARAR EL CODIGO PARA SEPARAR LA COLUMNA ANIDADA
expanded_records = []

for index,row in game_items.iterrows():
    user_id  = row['user_id']
    
    steam_id = row['steam_id']
    user_url = row['user_url']
    items_count = row["items_count"]
    items_list = row['items']

    for it in items_list:
        # diccionario del nuevo df expandido
        new_item = {    'steam_id' : steam_id ,
                        'user_url' : user_url,
                        "items_count": items_count,
                        # **it
                        'item_id'  : it['item_id'],
                        'item_name'  : it['item_name'],
                        'playtime_forever'  : it['playtime_forever'],
                        'playtime_2weeks'  : it['playtime_2weeks']
                    }
        expanded_records.append(new_item)

game_items2 = pd.DataFrame(expanded_records)

#PASAR EL ARCHIVO A CSV
game_items2.to_csv("game_items.csv", index=False)

TRABAJAR EL TERCER ARCHIVO

#ELIMINAR LOS NAN
steam.dropna(how="all", inplace=True)

#PASAR LOS ARCHIVOS A CSV
steam.to_csv("steam_games.csv", index=False)

steam
