==> 2. NPL.ipynb <==
#importar las librerias
import pandas as pd
import numpy as np

#importar la base de datos para trabjarla
game_reviews= pd.read_csv("game_reviews.csv", index_col= False)

game_reviews["review"].info

#limpiar la base de datos de puntuacion
game_reviews["review"]=game_reviews["review"].str.replace(r"[\"\'\|\?\=\.\@\#\*\,\.\!\/\:\-\‚ô•\<\>√â\+\D]", '',regex=True)

game_reviews.info()

#importar la libreria para hacer el analisis de sentimientos
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
nltk.download("vader_lexicon")

#importar la librerias para el analisis de sentimientos
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer
nltk.download("vader_lexicon")
sid=SentimentIntensityAnalyzer()


#aplicar el analisis de sentimientos a la columna reviews
game_reviews["Sentiment"]=game_reviews["review"].astype(str).apply(lambda i: sid.polarity_scores(i)["compound"])


#reemplzar los reviews vacios por nan
game_reviews["review"].replace("", np.nan, inplace=True)
#game_reviews.fillna(np.nan, inplace=True)
#game_reviews["review"].fillna(np.nan,inplace=True)

#PASAR LA NUMERACION QUE SID PASA A LA NUMERACION QUE SE REQUIERE PARA EL ANALISIS
Sentiment_Result=[]

for i in range(len(game_reviews)):
    if game_reviews["Sentiment"].iloc[i]==0:
        Sentiment_Result.append(1)
    elif game_reviews["Sentiment"].iloc[i]<0:
        Sentiment_Result.append(0)
    elif game_reviews["Sentiment"].iloc[i]>0:
        Sentiment_Result.append(2)

#CREAR LA COLUMNA DE RESULTADO
game_reviews["Sentiment_Result"]=pd.DataFrame(data=Sentiment_Result, index=None)

#Eliminar las columnas que no sirven del dataframe
game_reviews.drop(columns=["last_edited", "funny"], inplace=True)

game_reviews.to_csv("game_reviews2.csv")

"""from langdetect import detect, LangDetectException

# Función para detectar el idioma de un texto
def detectar_idioma(texto):
    if len(texto) >= 3:  # Establece un límite mínimo de longitud del texto
        try:
            return detect(str(texto))
        except LangDetectException:
            pass
    return None

# Aplica la detección de idioma solo a la columna "mensaje"
df6["idioma"] = df6["mensaje"].map(detectar_idioma)"""
