==> 4.Funciones-2.ipynb <==
def countreviews( YYYY-MM-DD y YYYY-MM-DD : str ): Cantidad de usuarios que realizaron reviews entre las fechas dadas  y, el porcentaje de recomendación de los mismos en base a reviews.recommend.

#IMPORTAR LIBRERIAS
import pandas as pd
import json


#USAR SOLO LAS COLUMNAS QUE SIRVEN
df_counter=pd.read_csv("game_reviews2.csv", usecols=["user_id","posted","recommend"])

#DIVIDIR LAS COLUMNAS DE FECHA PARA PODER LIMPIAR LOS DATOS
df1=df_counter["posted"].str.split(expand=True)

#ESCOGER LAS COLUMNAS UTILES
df1=df1.iloc[:,1:4]

#UNIR COLUMNAS QUE FUERON DIVIDIDAS
df_counter=pd.concat([df_counter,df1], axis='columns').drop(columns=["posted"])

#RENOMBRAS LAS COLUMNAS
df_counter.rename(columns={1:"Month",2:"Day",3:"Year"}, inplace=True)

df_counter

#LIMPIAR LA PUNTUACION EN LAS COLUMNAS PARA LUEGO PODER UNIRLAS
import string
df_counter["Day"] = df_counter["Day"].str.replace('[{}]'.format(string.punctuation), '',regex=True)
df_counter["Year"] = df_counter["Year"].str.replace('[{}]'.format(string.punctuation), '',regex=True)

#REEMPLAZAR LOS AÑOS QUE NO ESTAN CON LA MODA
df_counter['Year'].fillna(df_counter['Year'].mode()[0], inplace=True)

#REEMPLAZAR LA COLUMNA DE MESES CON LOS MESES NUMERICOS
df_counter["Month"].replace({"January":"01","February":"02","March":"03","April":"04","May":"05",
     "June":"06","July":"07","August":"08","September":"09","October":"10","November":"11","December":"12"},inplace=True)

#CREAR UNA COLUMNA DE FECHA DONDE ESTAN JUNTAS LAS FILAS DE DIA, MES, AÑO
df_counter["Fecha"]= df_counter["Year"]+"-"+df_counter["Month"]+"-"+df_counter["Day"]

#CONVERTIR LA COLUMNA DE FECHA EN TIPO STRING
df_counter["Fecha"]=df_counter["Fecha"].astype(str)

#HACER LA FUNCION QUE CONVIERTE LA FECHA EN TIPO DATETIME
def time(df):
    for index,item in enumerate(df["Fecha"]):
        try:
            datetime.strptime(item, "%Y-%m-%d").date()
        
        except:
            print(index)
            print(item)

#REEMPLAZAR LOS VALORES DIFERENTES COMO LA DEL 29 DE FEBRERO
df_counter["Fecha"].replace({"2014-02-29":"2014-02-28"}, inplace=True)

#ENCONTRAR LA MODA PARA REEMPLAZARLO EN LA FECHA
df_counter["Fecha"].mode()

#REEMPLAZAR LA MODA EN LAS FILAS DONDE NO HAY VALORES
df_counter["Fecha"].replace({"nan":"2014-06-21"}, inplace=True)

#APLICAR LA FUNCION A LA DF_COUNTER
time(df_counter)

#ELIMINAR LAS COLUMNAS QUE NO SIRVEN 
df_counter.drop(columns=["Month","Day","Year"],inplace=True)

df_counter.to_csv("df_counter_func_2.csv",index=False)

df_counter.info()

#CONTAR LOS REVIEWS POR FECHAS
(df_counter[df_counter["Fecha"]>"2011-11-5"]["recommend"]).nunique()

df_counter["user_id"].nunique()

#CONTAR LOS USER POR FECHAS
df_counter[(df_counter["Fecha"]>"2011-11-5")& (df_counter["Fecha"]<"2012-01-30")]["user_id"].nunique()

#CONTAR LAS REVIEWS POR FECHAS
recommend=df_counter[(df_counter["Fecha"]>"2011-11-5")& (df_counter["Fecha"]<"2014-07-8")]["recommend"]

recommend.value_counts()[True]

#ENCONTRAR EL PORCENTAJE DE RECOMENDACIONES 
porce_recom=(recommend.value_counts()[True])/len(recommend)

#FUNCION DE RECOMENDACION
def countreviews( fecha1,fecha2 : str ):
    df_counter=pd.read_csv("df_counter_func_2.csv")
    cantidad_usu_rese=df_counter[(df_counter["Fecha"]>fecha1)& (df_counter["Fecha"]<fecha2)]["user_id"].nunique()
    recommend=df_counter[(df_counter["Fecha"]>fecha1)& (df_counter["Fecha"]<fecha2)]["recommend"]
    true_count = recommend.value_counts().get(True, 0)  
    porce_recom = true_count / len(recommend) if len(recommend) > 0 else 0  

    response = {'cantidad_usuarios':cantidad_usu_rese, 'porcentaje_recomendacion':round(porce_recom,2)}
    
    response = {'cantidad_usuarios':cantidad_usu_rese, 'porcentaje_recomendacion':round(porce_recom,2)}
    return response

countreviews("2011-01-01","2014-01-30")
