==> 6.Funciones-4.ipynb <==
def userforgenre( género : str ): Top 5 de usuarios con más horas de juego en el género dado, con su URL (del user) y user_id.

import pandas as pd
import numpy as np
import ast

# Crear una tabla para las las horas jugadas por genero de cada usuario
# Leer dataframe solo con columnas a importar
columnas = ['id','genres'] 
df = pd.read_csv('steam_games.csv',usecols=columnas)
df.rename(columns={'id':'item_id'},inplace=True)
# remover los items que no tienen un genero definido
df = df.dropna(subset=['genres']) 

# Funcion que cambia de string a formato list los objetos
def tolist(lst):
    if isinstance(lst, str) :
        lst = lst.replace("0's",'0s')
        lst =  ast.literal_eval(lst)

    return list(lst)
# Mapear la columna con la funcion creada
df['genres'] = df['genres'].apply(lambda x : tolist(x))

# Función para crear las columnas binarias de los géneros
def create_genre_columns(dataframe):
    genres_list = set()

    # se crea el set con los valores DISTINCT de los géneros
    for genres in dataframe['genres']:
        if isinstance(genres, list):
            genres_list.update(genres)
    
    # crea una columna por cada genre en genre_list 
    # se recorre la columna genre imputando 1 cuando existe ese género 
    # en la fila analizada
    for genre in genres_list:
        dataframe[genre] = dataframe['genres'].apply(lambda x: 1 if genre in x else 0)
    
    return dataframe

# Aplicar la función
df_binario = create_genre_columns(df)
# Eliminar la columna genres
df_binario.drop('genres',axis=1,inplace=True)

#Importar el csv de los items por usuario
columnas = ['steam_id','user_url','item_id','playtime_forever'] # Columnas a importar
df_users = pd.read_csv('game_items.csv',usecols=columnas)
# Merge con el df_binario recientemente modificado
df_users = df_users.merge(df_binario,on='item_id',how='left')
# Imputar nulos con cero 0
df_users.fillna(0,inplace=True)
df_users.rename(columns={"steam_id":"user_id"}, inplace=True)

df_users

# Agrupar por usuario y calcular la suma total de tiempo jugado
col_sum = [ 'Indie','Racing','Sports','Audio Production','Animation &amp; Modeling',
            'Video Production','Software Training','Casual','Simulation','Utilities',
            'Photo Editing','Early Access','Accounting','Education',
            'Massively Multiplayer','Action','Web Publishing',
            'Adventure','Strategy','RPG','Design &amp; Illustration','Free to Play']
for col_name in col_sum:
    df_users[col_name] = df_users['playtime_forever'] * df_users[col_name]
df_users_group = df_users.groupby(['user_id','user_url'])[col_sum].agg('sum').reset_index()
df_users_group.columns = df_users_group.columns.str.lower()
df_users_group.head(3)

# Guardar el df en un csv que consumirá la API
df_users_group.to_csv('df-funciones-4.csv',index=False)

# Función que devuelve el top 5 de usuarios que más jugaron un género dado
def genre_top5_users(text : str):
    text = text.lower().strip()
    df1 = pd.read_csv('df-funciones-4.csv')
    if text not in df1.columns:
        return "Genre not found"
    
    top5 = df1.sort_values(by=text,ascending=False).head(5).reset_index()
    response = []
    for i in range (0 , 5):
        user_id = top5['user_id'].loc[i]
        user_url = top5['user_url'].loc[i]
        item_dict = {"user_id":user_id,
                     "user_url":user_url}
        response.append(item_dict)
    return response

genre_top5_users("Sports")
