==> 7.Funciones-5.ipynb <==
def developer( desarrollador : str ): Cantidad de items y porcentaje de contenido Free por año según empresa desarrolladora. Ejemplo de salida:

import pandas as pd
import numpy as np
import re

# Leer dataframe con columnas necesarias
columns = ['id', 'price', 'developer','title', 'release_date']
games = pd.read_csv('steam_games.csv', usecols=columns).rename(columns={'id':'item_id'})
games.sample(10)

# Función para corregir el precio, en el que aparecen valores numéricos y en formato string
def fix_price(df):

    errors_list = []
    for i in df['price']:
        try:
            float(i)
        except:
            errors_list.append(i)

    errors = set(errors_list)
    #uniques_not_free = ['Starting at $499.00', 'Starting at $449.00']
    df['price_fixed'] = df['price'].apply(lambda x: 0 if x in errors 
                                                        else 499.0 if x=='Starting at $499.00'
                                                        else 449.0 if x=='Starting at $449.00'
                                                        else x)
    df['price_fixed'] = df['price_fixed'].astype(float)
    return df

# Función para convertir la columna fecha en dtype datetime
def safe_date_convert(df, date_column):
    def convert(x):
        try:
            return pd.to_datetime(x)
        except ValueError:
            # Try to find a year pattern in the string
            year_pattern = re.search(r'\b\d{4}\b', x)
            if year_pattern:
                # Convert the found year into a datetime format
                return pd.to_datetime(year_pattern.group(0), format='%Y')
            else:
                # If no year pattern is found, return NaT
                return pd.NaT

    df['date_fixed'] = df[date_column].astype(str).apply(convert)
    return df

# Transformar la columna price a numérico
games=fix_price(games)
# Transformas la columna de fechas a dtype datetime
games = safe_date_convert(games, 'release_date')

# Obtener el año de lanzamiento
games['release_year'] = games['date_fixed'].dt.year
# Eliminar columnas innecesarias
games.drop(columns=['release_date','price','date_fixed'],inplace=True)
games.drop(columns=['title'], inplace=True)
games.head()

# Imputar los nulos de precio como 0 y los na de release_year por la moda
# Los nulos de desarrollador no puedo imputarlos, así que los elimino.
games.dropna(subset=['developer'], inplace=True)
games['price_fixed'].fillna(0,inplace=True)
games['release_year'].fillna(games['release_year'].mode().iloc[0],inplace=True)

# Verificar y eliminar duplicado
games.duplicated().sum()
games.drop_duplicates(inplace=True)

# Agrupar por desarrollador para calcular el numero de videojuegos lanzados por año
dev = games.groupby(['developer', 'release_year'], as_index=False).agg({'item_id':'count'}).rename(columns={'item_id':'item_count'})
# Calcular número de juegos gratis
free_games = games[games['price_fixed']==0].groupby(['developer','release_year'], as_index=False)['item_id'].agg('count').rename(columns={'item_id':'free'})
# Merge de los dos df (left join)
dev = pd.merge(dev, free_games, on=['developer', 'release_year'], how='left')
# Imputar nulos de la columna free con 0
dev['free'].fillna(0,inplace=True)
# Calcular porcentaje de videojuegos gratis por desarrollador por año
dev['porcentaje_free'] = dev['free'] / dev['item_count'] * 100
# Sample
dev.sample(10)

# Convertir el año a entero
dev['release_year'] = dev['release_year'].astype(int)
# Convertir los nombres de desarrolladores a minúscula y eliminar espacios en blanco
dev['developer'] = dev['developer'].str.strip().str.lower()
# Redondear porcentaje de videojuegos gratis
dev['porcentaje_free'] = dev['porcentaje_free'].round(2)
# Eliminar la columna free
dev.drop(columns=['free'], inplace=True)
dev.sample(10)

# Guardar el df final a un csv que consumirá la API
dev.to_csv('df-funcion-5.csv',index=False)

# Función que retorna la cantidad de videojuegos lanzados en un año dado
# y el porcentaje de videojuegos gratis
def developer(desarrollador : str):
    df = pd.read_csv('df-funcion-5.csv')
    desarrollador = desarrollador.strip().lower()
    if df['developer'].str.contains(desarrollador).any():
        data = df[df['developer'] == desarrollador]
        data = data.sort_values('release_year', ascending=False)
        response = data[['release_year','item_count','porcentaje_free']].to_dict(orient='records')
        return response
    else:
        return 'Error: Developer not found'

# Probar la función
developer('capcom')
