==> 8.Funciones-6.ipynb <==
def sentiment_analysis( año : int ): Según el año de lanzamiento, se devuelve una lista con la cantidad de registros de reseñas de usuarios que se encuentren categorizados con un análisis de sentimiento.

import pandas as pd
import numpy as np

user_revs=pd.read_csv("game_reviews2.csv", usecols=['user_id','item_id', 'Sentiment_Result'])

# Revisar tipo de datos y cantidad de nulos
print(user_revs.info())
print(user_revs.isna().sum())
# Eliminar nulos 
user_revs.dropna(inplace=True)

# Leer dataframe de videojuegos con las columnas necesarias
columns = ['id', 'release_date']
games = pd.read_csv('steam_games.csv', usecols=columns).rename(columns={'id':'item_id'})
games.sample(15)

# Revisar tipos de datos y cantidad de nulos 
print(games.info())
print(games.isna().sum())
games.dropna(inplace=True)

# Hacer merge (left join) de los dos df
df_sa = user_revs.merge(games, on='item_id', how='left')
# Imputar nulos en la columna fecha con la moda
df_sa.fillna(df_sa['release_date'].mode()[0], inplace=True)
df_sa.sample(10)

# Convertir la columna release_date a dtype datetime
df_sa = safe_date_convert(df_sa, 'release_date')
df_sa.info()

print(df_sa['date_fixed'].dtype)
# Verificar si hay valores nulos o faltantes
print(df_sa['date_fixed'].isnull().sum())
# Convertir la columna "date_fixed" al tipo de dato "datetime" si es necesario
df_sa['date_fixed'] = pd.to_datetime(df_sa['date_fixed'], format='%Y-%m-%d', errors='coerce')

# Obtener el año de lanzamiento
df_sa['year_released'] = df_sa['date_fixed'].dt.year
# Quedarse solo con las columnas que se necesitan
df_sa = df_sa[['Sentiment_Result','year_released']]


df_sa.info()

# Calcular los porcentajes de reviews positivos, negativos y neutros
sentiment_analysis = pd.DataFrame()
sentiment_analysis['Negative'] = df_sa[df_sa['Sentiment_Result']==0].groupby('year_released').agg('count')
sentiment_analysis['Neutral'] = df_sa[df_sa['Sentiment_Result']==1].groupby('year_released').agg('count')
sentiment_analysis['Positive'] = df_sa[df_sa['Sentiment_Result']==2].groupby('year_released').agg('count')


# Guardar el df final en un csv que consumirá la API
sentiment_analysis.to_csv('sentiment_analysis.csv')

# Función que retorna el porcentaje de reviews negativos, positivos
# y neutros en un año especificado
def sentiment_analysis(año : str):
    df = pd.read_csv('dataquery/sentiment_analysis.csv')
    df['year_released'] = df['year_released'].astype(str)
    if df['year_released'].str.contains(año).any():
        return df[df['year_released'] == año].to_json(orient='records')
    else:
        return 'Year not found'
