==> 10.ML.ipynb <==
K-neighbors es el modelo mas adecuado para este sistema de recomendaciones. Este modelo puede definir istintas medidas de similaridad como distancia euclidiana o cosine similarity.

#Tranformar el dataset

#importar las librerias y leer el archivo
import pandas as pd

games = pd.read_csv('steam_games.csv')

#hacerle cambios a la columna de release date para que quede con formato de fecha, 
#crear una columna de año y rellenar losvalores faltantes 
games['release_date'] = pd.to_datetime(games['release_date'], errors='coerce')
games['release_year'] = games['release_date'].dt.year
games['release_year'] = games['release_year'].fillna(0).astype(int)
games['release_year'] = games['release_year'].astype('str').replace('0', '')

games.fillna('', inplace=True)

games.drop(games[games['title'] == ""].index, inplace=True)
games.shape

# limpiar la columna developer para que no tenga contenga caracteres inusuales'
games['developer'] = games['developer'].str.replace(r'\s+', '-', regex=True)
games['developer'] = games['developer'].str.replace(r'---', "-")

import re
import ast

# función de limpieza de las columnas
#la primera parte del codigo funciona para volver a forma de lista un conjunto de datos, esto se usara en las columnas anidadas
def tolist(lst):
    text=""
    if isinstance(lst, str) and len(lst) > 0 and '[' in lst:
        lst = lst.lower()
        lst = lst.replace("0's",'0s').replace(';','').replace('&','').replace("'em","em")
        try:
            lst =  ast.literal_eval(lst)
        except (SyntaxError, ValueError):
            return lst.lower()
        
    if isinstance(lst,list):
        for item in lst:
            while "  " in item:
                item = item.replace("  "," ")
            item = item.replace(" ", "-")
            text = text + " " + item 
            text = text.replace('"','')
        return text.strip().lower()

    return lst

# aplicamos la funcion a cada columna anidada
games['specs'] = games['specs'].apply(lambda x : tolist(x))
games['genres'] = games['genres'].apply(lambda x : tolist(x))
games['tags'] = games['tags'].apply(lambda x : tolist(x))

# Combine relevant columns into a single 'profile'
games['profile'] = games['genres'] + " " + games['tags'] + " " + games['specs'] + " " + games['developer'] + " " + games['release_year']

# Dataset listo para usarse en el modelo
games[['id','title','profile']].to_csv("ml_dataset.csv",index=False)

#aplicar el modelo K-neighbors

#importar la librerias para trabajar el modelo, es importante la libreria de vectorizar para hacer vectores
#con las palabras y la libreria del modelo K-neighbors
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.neighbors import NearestNeighbors

df = pd.read_csv("ml_dataset.csv")

# se crea un vector con todos los datos de la columna profile para que se usen en el modelo
tfidf = TfidfVectorizer(stop_words='english',max_features=5000)
tfidf_matrix = tfidf.fit_transform(df['profile'])

# usar la libreria K-neighbors para encontrar los uegos similares
knn = NearestNeighbors(n_neighbors=6, metric='cosine', algorithm='brute')  # Se coloca 6 pues tiene que considerar el juego del que se tiene referencia
knn.fit(tfidf_matrix)

# Crear función que crea las recomendaciones
def recommend(game_id):
    #agrega los indices de juegos iguales al game_id y devuelve el primero que encuentre
    try:
        idx = df.index[df['id'] == game_id].tolist()[0]
        #obitnene el nombre del juego con el indice encontrado en la linea superior
        name = df['title'].loc[idx]
        
        values , indices = knn.kneighbors(tfidf_matrix[idx])
        
        # se excluye el primer valor pues es el id que se puso en la funcion
        game_indices = indices[0][1:]
        result = df['id'].iloc[game_indices]
    except:
        return [0]
    return list(result)

# test
recommend(47810)

# un df sin la columna profile
df1 = df[['id','title']]

# función que crea la columna con las recomendaciones de cada juego
df1['recommends'] = df1['id'].apply(lambda x  : recommend(x))

# Dataset para usar en la API listo 
df1['id'].fillna(0,inplace=True)
df1.to_csv("modelo.csv",index=False)
df1.info()

df1

Funcion de recomendacion

# Función del endpoint
import pandas as pd
import ast

def recomendacion_juego( game_id : int ):
    # carga de archivos
    df = pd.read_csv("modelo.csv")

    if game_id not in df['id'].values:
        return 'error'

    # Obtiene la lista de recomendaciones
    result = df[df['id'] == game_id]['recommends'].iloc[0]

    try:
        result = ast.literal_eval(result)
    except (SyntaxError, ValueError):
        print('No se puede convertir a lista')
        return 'error'

    response =[]
    item_buscado = df[df['id'] == game_id].iloc[0]
    
    for item_id in result:
        # Obtiene la información del juego recomendado
        item_info = df[df['id'] == item_id].iloc[0]  

        response.append({'game_id': int(item_info['id']), 
                         'title': item_info['title']})
    dict_respuesta = {'titulo_buscado': item_buscado['title'], 'results':response}  

    return dict_respuesta

# test
recomendacion_juego(47810)
